pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wordpress-db-explorer"
version = "1.0.0"
description = "A powerful command-line tool for exploring, searching, and exporting data from WordPress databases"
authors = [
    {name = "WordPress DB Explorer Team"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "SQLAlchemy",
    "rich",
    "inquirer",
    "python-dotenv",
    "pymysql",
]

[project.optional-dependencies]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--color=yes",
    "--strict-markers",
]
markers = [
    "unit: Unit tests that test individual functions and components",
    "integration: Integration tests that test component interactions",
    "slow: Slow running tests that may take longer to execute",
    "database: Tests that require database connection",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
